requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chess==1.11.2
contourpy==1.3.2
curl_cffi==0.10.0
cycler==0.12.1
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
executing==2.2.0
FLAML==2.3.4
fonttools==4.58.0
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==8.36.0
jedi==0.19.2
jiter==0.9.0
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
multitasking==0.0.11
numpy==1.26.4
openai==1.77.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
peewee==3.18.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyautogen==0.2.25
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
termcolor==3.1.0
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
yfinance==0.2.59
